+++ sdk/storage/Azure.Storage.Blobs/tests/BlockBlobClientTests.cs
@@ -2811,18 +2811,5 @@ namespace Azure.Storage.Blobs.Test
             public int CommittedCount { get; set; }
             public int UncommittedCount { get; set; }
         }
-
-        private async Task<BlockBlobClient> GetNewBlobClient(BlobContainerClient container, string blobName = default)
-        {
-            blobName ??= GetNewBlobName();
-            BlockBlobClient blob = InstrumentClient(container.GetBlockBlobClient(blobName));
-            var data = GetRandomBuffer(Constants.KB);
-
-            using (var stream = new MemoryStream(data))
-            {
-                await blob.UploadAsync(stream);
-            }
-            return blob;
-        }
     }
 }