+++ sdk/storage/Azure.Storage.Files.DataLake/CHANGELOG.md
@@ -6,6 +6,8 @@
 - Added DataLakeDirectoryClient.GetPaths().
 - Fixed bug where Data Lake SDK coudn't handle SASs with start and expiry time in format other than yyyy-MM-ddTHH:mm:ssZ.
 - Added ability to set Position on streams created with DataLakeFileClient.OpenRead().
+- Added CanGenerateSasUri property and GenerateSasUri() to DataLakePathClient, DataLakeFileClient, DataLakeDirectoryClient and DataLakeFileSystemClient.
+- Added CanGenerateAccountSasUri property and GenerateAccountSasUri() to DataLakeServiceClient.
 - Restored single upload threshold for parallel uploads from 5 TB to 256 MB.
 
 ## 12.5.0-preview.1 (2020-09-30)