+++ sdk/storage/Azure.Storage.Queues/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 12.5.0-preview.2 (Unreleased)
 - Fixed a bug where QueueServiceClient.SetProperties and QueueService.GetProperties where the creating/parsing XML Service Queue Properties CorsRules incorrectly causing Invalid XML Errors
 - Fixed bug where Queues SDK coudn't handle SASs with start and expiry time in format other than yyyy-MM-ddTHH:mm:ssZ.
+- Added CanGenerateSasUri property, GenerateSasUri() to QueueClient.
+- Added CanGenerateAccountSasUri property, GenerateAccountSasUri() to QueueServiceClient.
 
 ## 12.5.0-preview.1 (2020-09-30)
 - This preview contains bug fixes to improve quality.